fix: rename files to .ts extension
